Implemented Ragdoll on death.
The active ragdoll setup will be harder than I thought and because it's a precision platformer/hack-and-slash it'll be better to just make you only ragdoll when you die. So I'll leave it as it is now, and then work on the design of that in the future.

TODO:
Implement the code for doing damage, make sure you really understand how it'll work. Maybe make an interface for it to make life easier.