Implement Collision sensor
We now use a "CollisionSensor" with our weapon to check if we're hitting an object. We can define the size and shape of our collision sensor and it'll have a wedge shape to it. This is much better than the old way of checking collisions.

I'm going to begin modifying the player movement some more so that it's more like a platformer.

TODO:
Finish player code, then move on to bosses.