Started getting the statemachine for the boss working.
Currently the boss will follow the player around until the player is in attack range, and when the boss moves to the player's position, they don't constantly track the player they instead move to the position they were at when they started the move state.

TODO:
Change colors depending on state and begin implementing attack class and scripts.