Update PlayerTestScene.Unity
Fixed an issue where the UI wasn't scaling properly with the screen size which caused the LowRes effect to not work.

Setting canvas scaling to "Match width or height" and then setting it to 0.5 between width and height then setting the lowRes Image object to stretch and fill the canvas works.

TODO:

Finish the boss attacking and make an "Action class" for the boss to do actions instead of the "Attack" Class.