Update BossController.cs and DashAwayMove.cs
I implemented some changes to the MoveToPosition coroutine so that it has a speed and accuracy parameter which means when the boss is within a certain range they'll stop moving once reaching that range and they'll move at the speed passed. By default the speed is 10 and the targetAccuracy is 5. This makes it so when the boss is decently close to the player they stop moving, switch to attack, and then dodge away at a considerable speed.

TODO:
Make the boss rotate to face the player when moving.