Update CollisionSensor.cs
Fixed a bug where the collisionSensor would throw an error if a boss died and then the collisionSensor attacked a different damageable. This is now fixed and there is no error. It was because the transform was null and wasn't removed from the transforms list. It is now removed correctly.