Update BossController.cs and Weapon.cs
The boss controller now moves towards the player and stops as soon as they are in attack range, then attacks, and dodges away.

The weapon.cs script now properly uses the "AttackDistance" parameter. It updates in edit mode as well.

TODO:
Make the player become stunned when they are attacked and make the boss idle for a moment before they attack the player.

Also change the action class to use a coroutine instead that way we can make the attacks and actions work easier.