Update BossController.cs
Collision avoidance is kind of working, I'll have to tweak it more and work on it more but it's getting there. The sphere casts are what make our avoidance possible, but sometimes there will be false collisions found which is odd. I need to modify the values used in calculating collisions, focus on the force for collisions as that's why it sometimes gets stuck, but we need a high value there.