BossController.cs
Finalized the parameters for collision avoidance for the boss.

Going to move to implementing the different attack types for the weapon and then also implementing an actual first boss.

TODO:
Implement alt and special attacks for weapons

Add a cooldown to the dashing for the player.

Implement the crocodile boss to test the gameplay more.