Update BossController.cs
I cleaned up some of my code for the steering algorithm by converting each rule into a function. Then I implemented the collision avoidance and steering algorithm in the MoveToPosition Coroutine which means that now the dash attack has collision avoidance built into it.

TODO:
Figure out why when the boss is hit and it can attack the player, instead of dodging away like it should it ignores the attack and damages the player.