Update BossController.cs
I fixed an issue where the iframes coroutine didn't wait a frame after switching to the stun state so any checks done to stop a coroutine if the state changes wouldn't get a chance to check. I made the coroutine wait until the end of the frame before fully executing it's code so that fixed the issue.

The boss is fairly fun to fight and properly avoids obstacles now.

TODO:
Start working on making an actual new boss.