Update PlayerTestScene.unity, UIManager.cs, GameUICanvas.prefab
Started implementing chris' UI into the PlayerTestScene and updated his code so that the cursor is unlocked and visible when entering the pause menu or quitting to the main menu.

TODO: Link the UI in the hud to the player to display their health and other data, and also link the boss health bar to the currently active boss in the scene. Make the boss tell the UI to turn on the health bar and to update it from the boss.