Update UIManager.cs and Implement BossUIManager.cs
The boss's health bar now matches their health, and their name is shown above the bar now.

When the boss starts to exist it shows the bossUI and when it dies the bossUI is hidden.

TODO: update the player to have a cooldown of some kind on the dash and display the dash, and weapon cooldowns in the UI.