Update PlayerUIManager.cs & Player.cs & BossController.cs
Implemented the caps to the Player and the UI for the player.

1 cap is rewarded by the base boss.
3 coins are rewarded by the base boss.

Implemented a cooldown for the dash which might be too long right now, I may want to lower it.

The weapon indicator greys out when it is on cooldown and the dash indicator greys out when it is on cooldown.

TODO: Figure out why the boss always dashes back to the same position when using the DashAwayMove and then figure out how to make it so it dashes away to the position closest to the desired position that is unobstructed. That way it doesn't get stuck.

For now, the UI is properly implemented.