Update DashAwayMove.cs
I fixed an issue in the dash away move where sometimes the boss would target a position it couldn't reach on the y axis so it would move to a position below or above that point and get stuck in the move coroutine. This should no longer occur, but if it does it's due to the y component being modified and not changed back to the bosses normal y position.

We don't modify their y position right now.

TODO: Make sure the UI is implemented correctly for now and then merge this branch to main.