Implemented Squid boss and fixed attack animation issue
There was an issue where when the attack animation ends it wouldn't wait before moving onto the dash away animation and now it does properly wait for the attack animation to end before moving onto the dash away animation.

I'm going to change how the squid boss works so that it uses 2 different attacks, one where it dashes at the player like a torpedo and one where it just hits you with a tentacle.

I need to make a bubble particle effect that plays when it does the dash attack towards the player.