Implemented the gachamachine.
TODO:

I need to implement the actual visuals for the gacha machine.

I may want to change the CapsuleDrop to a be a scriptable object so that we don't need a prefab for every one. For now, it'll work fine.

Next I'm going to finish working on the blender animation for the capsule and just quickly finish it up so that I can get it implemented in the game and finish up the visuals.