Started implementing spin attack and fix bug in smash attack.
I fixed a bug in the smash attack where on the alt attack it would randomly send the player flying because gravity was set to 9.81f prematurely and adding a check for didLand with the !jumping check fixed it.

TODO: Finish tuning the slam attack and finish deciding how to implement the floating effect caused by the spin attack.