Implemented Special attack and redid attack controls
Left trigger is dash, Button West is attack, Button North is alt-attack, button east is special attack. Button south is jump. Clicking the left stick in is interact.

TODO: Figure out how to make the special attack end earlier because for some reason it has a bit longer before you can attack again and I want the player to start an attack at the end of it so they can do combos like in castle crashers.