Implemented Stunning the boss so combos are possible.
Bosses are currently slowly pushed into the air with every attack and frozen for a small amount of time which currently is good to start having actual combos.

I need to add a grounded check for the boss in the next commit and change how I check if we should unlock the Y axis in ApplyFinalMovements for the boss.

This also opens opportunities for landing effects for the boss.

I also may need to make the bosses heavier so the player can't push them when he's moving.

Or I should make the player collider ignore collisions with bosses.