Added ground check for boss and bug fix for weapons
The collision sensor layer mask wasn't working how I intended so I just changed it so there's a check if the object we hit is a parent of the weapon at all and if it is then we ignore it. This was the simplest way to fix the issue.

TODO: Add some code to the weapon that allows to damage enemies in a sphere and disc instead of just using the collision sensor. You overcomplicated that a bit.