Got the spin attack collision working correctly.
Changed how the collision sensor works so now I rotate the collision sensor with the spin attack and actually register when and where the hit occurs so that the spin attack works.

I'm gonig to just play around a bit in the game and see how it feels to fight bosses then come back to working on this before I pull from main and merge my changes.