Implemented the T-Rex boss for now
TODO: The t-rex boss needs to have a tail spin attack and a jump and stomp attack but for now it won't because it'll take too long to work on. The boss is implemented in a basic manner and can spawn from the machine.

We can make him more complex in the future.