Fixed the bug with being able to get multiple capsules and the bug with the weapon pickup radius.
TODO: merge this back to main now that everything is fixed.

I just added a check to ensure the capsule and the item it drops don't exists and it isn't in the dispense animation before starting to dispense another item. That's what fixed the doubling up issue with the gashapon machine.

I also fixed the weapon pickup radius issue because at some point I added a check to onTriggerEnter to check for the weapon but we aren't supposed to check for picking it up there. We only check in onTriggerStay.