Coded a test WhiteMaterialToggle.cs
Pressing V toggles the UFO grabber to be white and pressing it again restores it's color.

This is just a test to show that swapping materials using this script works and restoring them works, including if it has a random seed assigned it'll be reassigned as well.

This is all handled by a boolean in the script called "isWhite" which has a property getter and setter which will call the setMaterialsWhite and SetMaterialsDefault on it's own.

TODO: When a gachamachine is locked, make sure it is set to white.

TODO: Implement having a gachamachine be locked so it can't be used.