Implemented locking environmental objects using a script.
The UFOGrabber now is locked by default and won't dispense anything when hit until it is unlocked.

Press V to toggle between locked and unlocked for the object.

TODO:
Code a new drop that when dropped simply calls a new custom public method that has the name of the gachamachine in it's parameters and unlocks that specific machine.