Implemented UFO Grabber as a drop
The UFO grabber can now drop from the base machine and it looks great!

it is also saved so that once it is collected if you die, or exit the game and reload, the UFO grabber will be unlocked on start!

TODO:
Change the drop chance for the UFO grabber to be lower than everything else and then start implementing the UFO boss.

First though, make sure the UFO grabber works and merge it to main.