Fix: LaunchPlayer and added proper alt attack for the bomb
LaunchPlayer would sometimes launch the player and the gravity would get set to 9.81 after it was already changed to 60 by the launch calculation and I fixed this by adding a bool which indicates when the player is launched so now the launch is always consistent.

I also added an alt attack to the bomb which propels the player upwards using launch by immediately detonating a bomb at their feet.