Update PinkPlastic.mat & Bug fix for CollisionSensor.cs
The collision sensor wouldn't detect collisions for large hitboxes because the center was too far so I had to use the bounds and the extents to calculate if the positions of subdivided points within the bounding box fit into the arc, and if they do it counts as a hit.

This worked way better than expected and I was able to have fun fighting the UFO boss.

Honestly the UFO boss is the most fun so far.