Implemented Laser attack visuals and fixed post-processing visuals issue.
The laser attack happens but it doesn't currently deal any damage yet. Need to implement that.

I was able to fix the camera post processing issue by checking the post-processing box in the main camera prefab. This made it so I could finally see the bloom FX  on the Mechasaurus laser attack.

TODO:
Implement a hitbox for the Mechasaurus laser and make it deal damage.